+++ src/private/GetDuplicateFunctionGroup.ps1
@@ -1,0 +1,12 @@
+function Get-DuplicateFunctionGroup {
+    [CmdletBinding()]
+    param(
+        [Parameter(Mandatory)][System.Management.Automation.Language.FunctionDefinitionAst[]]$FunctionAst
+    )
+
+    return @(
+        $FunctionAst |
+            Group-Object -Property { ('' + $_.Name).ToLowerInvariant() } |
+            Where-Object { $_.Count -gt 1 }
+    )
+}