+++ src/public/InvokeMTBuild.ps1
@@ -5,6 +5,12 @@ function Invoke-MTBuild {
     $ErrorActionPreference = 'Stop'
     Reset-ProjectDist
     Build-Module
+
+    $data = Get-MTProjectInfo
+    if ($data.FailOnDuplicateFunctionNames) {
+        Assert-BuiltModuleHasNoDuplicateFunctionName -ProjectInfo $data
+    }
+
     Build-Manifest
     Build-Help
     Copy-ProjectResource